Support archive clone + move project setup logic from utils
